--- test/scholar/model_selection/k_fold_test.exs
@@ -1,5 +1,0 @@
-defmodule Scholar.ModelSelection.KFoldTest do
-  use Scholar.Case, async: true
-  alias Scholar.ModelSelection.KFold
-  doctest KFold
-end
