+++ test/scholar/model_selection/model_selection_test.exs
@@ -1,0 +1,5 @@
+defmodule Scholar.ModelSelectionTest do
+  use Scholar.Case, async: true
+  alias Scholar.ModelSelection
+  doctest ModelSelection
+end
